requirements.txt
jinja2
reportlab
python-dotenv
streamlit
google-genai